# api/openapi.yaml
openapi: 3.1.0
info:
  title: Data Space Policy Engine Service
  description: Standalone policy-as-code microservice. Evaluates user permissions
    based on role and institute using OPA.
  contact:
    name: HIRO-MicroDataCenters
    email: all-hiro@hiro-microdatacenters.nl
  license:
    name: MIT
    url: https://github.com/HIRO-MicroDataCenters-BV/ds-policy-engine/blob/main/LICENSE
  version: 0.1.0
paths:
...
  /api/v1/rules:
    get:
      tags:
      - Rules
      summary: List rules
      description: List all rules with pagination, filtering, and sorting
      operationId: list_rules_api_v1_rules_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          default: 10
          title: Page Size
      - name: sort_by
        in: query
        required: false
        schema:
          type: string
          default: created_at
          title: Sort By
      - name: sort_order
        in: query
        required: false
        schema:
          type: string
          default: desc
          title: Sort Order
      - name: role
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
      - name: enabled
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Rules
      summary: Create rule
      description: Create a new policy rule
      operationId: create_rule_api_v1_rules_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/rules/{rule_id}:
    get:
      tags:
      - Rules
      summary: Get rule
      description: Get a single rule by ID
      operationId: get_rule_api_v1_rules__rule_id__get
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          title: Rule Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Rules
      summary: Update rule
      description: Update an existing rule
      operationId: update_rule_api_v1_rules__rule_id__put
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          title: Rule Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Rules
      summary: Delete rule
      description: Delete a rule by ID
      operationId: delete_rule_api_v1_rules__rule_id__delete
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          title: Rule Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/rules/{rule_id}/toggle:
    patch:
      tags:
      - Rules
      summary: Toggle rule
      description: Enable or disable a rule
      operationId: toggle_rule_api_v1_rules__rule_id__toggle_patch
      parameters:
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          title: Rule Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/rules/import/json:
    post:
      tags:
      - Rules
      summary: Import rules from JSON
      description: Import rules from a previously exported JSON backup
      operationId: import_rules_api_v1_rules_import_json_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/rules/import/csv:
    post:
      tags:
      - Rules
      summary: Import rules from CSV
      description: Import rules from CSV content. Permissions should be semicolon-separated.
      operationId: import_rules_csv_api_v1_rules_import_csv_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CsvImportRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    CsvImportRequest:
      properties:
        csv_content:
          type: string
          title: Csv Content
        mode:
          type: string
          title: Mode
          default: merge
      type: object
      required:
      - csv_content
      title: CsvImportRequest
      description: Request body for CSV import.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ImportRequest:
      properties:
        rules:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Rules
        mode:
          type: string
          title: Mode
          default: merge
      type: object
      required:
      - rules
      title: ImportRequest
      description: Request body for rule import.
...
    RuleCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
          description: Rule display name
        description:
          type: string
          maxLength: 500
          title: Description
          default: ''
        role:
          type: string
          title: Role
          description: Target role for this rule
        institute:
          type: string
          title: Institute
          description: Institute this rule applies to
          default: ''
        permissions:
          items:
            type: string
          type: array
          title: Permissions
          description: Permissions granted by this rule. May be empty (the service
            surfaces such rules as a warning in the overview).
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      required:
      - name
      - role
      title: RuleCreate
      description: 'Schema for creating a new authorization rule.


        Defines the required and optional fields accepted when a client submits

        a request to create a rule. Each rule maps a role to a set of permissions.'
      examples:
      - description: Full permissions including cross-node capabilities
        enabled: true
        institute: local
        name: Catalog Owner Full Access
        permissions:
        - catalog:read
        - catalog:create
        - catalog:update
        - catalog:delete
        - decentralized-search:execute
        - federated-learning:execute
        role: catalog_owner
    RuleUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Description
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
        institute:
          anyOf:
          - type: string
          - type: 'null'
          title: Institute
        permissions:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Permissions
        enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
      type: object
      title: RuleUpdate
      description: 'Schema for partially updating an existing rule.


        All fields are optional; only the supplied fields are modified. Omitted

        fields retain their current values.'
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError